+++ python/revng/pypeline/utils/notification_broker.py
@@ -2,14 +2,14 @@
 # This file is distributed under the MIT License. See LICENSE.md for details.
 #
 
-import asyncio
 from abc import ABC, abstractmethod
+from asyncio import Queue
+from collections import defaultdict
 from typing import Protocol
 
-from starlette.websockets import WebSocket
-
-from revng.pypeline.storage.notification_queue import LOCAL_QUEUE
 from revng.pypeline.storage.storage_provider import ProjectID
+from revng.pypeline.utils import Locked
+from revng.pypeline.utils.logger import pypeline_logger
 
 
 class Stream(Protocol):
@@ -19,26 +19,13 @@ class Stream(Protocol):
     async def write(self, data: bytes) -> None: ...
 
 
-class WebSocketStream(Stream):
-    """Represents a WebSocket stream"""
-
-    def __init__(self, websocket: WebSocket):
-        self.websocket = websocket
-
-    async def read(self) -> bytes:
-        return await self.websocket.receive_bytes()
-
-    async def write(self, data: bytes) -> None:
-        await self.websocket.send_bytes(data)
-
-
 class NotificationSubscriber:
     """Represents a notification subscriber with its own message queue"""
 
-    def __init__(self, project_id: ProjectID | None, stream: Stream):
+    def __init__(self, project_id: ProjectID | None, stream: Stream, message_queue: Queue[bytes]):
         self.project_id = project_id
         self.stream = stream
-        self.message_queue: asyncio.Queue[bytes] = LOCAL_QUEUE.get_queue()
+        self.message_queue = message_queue
         """The broker will insert messages into this queue, and the subscriber will
         listen for them and insert them into the message queue."""
         self.is_active = True
@@ -71,12 +58,39 @@ class NotificationBroker(ABC):
         Pub/Sub.
     """
 
-    @abstractmethod
+    def __init__(self):
+        self.subscribers: Locked[dict[ProjectID | None, Locked[set[NotificationSubscriber]]]] = (
+            Locked(defaultdict(lambda: Locked(set())))
+        )
+
     async def subscribe(
         self, project_id: ProjectID | None, stream: Stream
     ) -> NotificationSubscriber:
         """Register to receive notifications for project \"project_id\" on the given stream"""
+        queue = await self.get_queue(project_id)
+        subscriber = NotificationSubscriber(project_id, stream, queue)
 
-    @abstractmethod
+        async with self.subscribers() as subscribers:
+            async with subscribers[project_id]() as project_subscribers:
+                project_subscribers.add(subscriber)
+
+        pypeline_logger.debug_log(f"Stream subscribed to project {project_id}")
+        return subscriber
+
     async def unsubscribe(self, subscriber: NotificationSubscriber):
         """Stop receiving notifications for project \"project_id\"."""
+        async with self.subscribers() as subscribers:
+            project_id = subscriber.project_id
+            async with subscribers[project_id]() as project_subscribers:
+                if project_id in subscribers:
+                    project_subscribers.discard(subscriber)
+
+        subscriber.close()
+        pypeline_logger.debug_log(f"Stream unsubscribed from project {subscriber.project_id}")
+
+    @abstractmethod
+    async def get_queue(self, project_id: ProjectID | None) -> Queue[bytes]:
+        """
+        Get the queue that the subscriber will receive messages in. This should
+        be exclusive to the subscriber.
+        """
